+++ src/top_items/git_bare_repos.rs
@@ -16,7 +16,7 @@ use crate::cache::*;
 use crate::tables::format_table;
 use crate::top_items::common::{dir_exists, FileDesc, Pair};
 
-use humansize::{file_size_opts, FileSize};
+use humansize::{FormatSize, DECIMAL};
 use rayon::prelude::*;
 use walkdir::WalkDir;
 
@@ -214,14 +214,9 @@ pub(crate) fn chkout_list_to_string(limit: u32, mut collections_vec: Vec<RepoInf
 
     for repoinfo in collections_vec.into_iter().take(limit as usize) {
         #[allow(clippy::integer_division)]
-        let average_size = (repoinfo.total_size / u64::from(repoinfo.counter))
-            .file_size(file_size_opts::DECIMAL)
-            .unwrap();
+        let average_size = (repoinfo.total_size / u64::from(repoinfo.counter)).format_size(DECIMAL);
 
-        let total_size = repoinfo
-            .total_size
-            .file_size(file_size_opts::DECIMAL)
-            .unwrap();
+        let total_size = repoinfo.total_size.format_size(DECIMAL);
 
         table_matrix.push(vec![
             repoinfo.name,
@@ -249,10 +244,7 @@ pub(crate) fn git_repos_bare_stats(
         output,
         "\nSummary of: {} ({} total)",
         path.display(),
-        bare_repos_cache
-            .total_size()
-            .file_size(file_size_opts::DECIMAL)
-            .unwrap()
+        bare_repos_cache.total_size().format_size(DECIMAL)
     )
     .unwrap();